Fix warning: implicit conversion changes signedness
visp/modules/sensor/src/force-torque/vpForceTorqueAtiNetFTSensor.cpp: In member function ‘bool vpForceTorqueAtiNetFTSensor::waitForNewData(unsigned int)’:
visp/modules/sensor/src/force-torque/vpForceTorqueAtiNetFTSensor.cpp:258:53: warning: conversion to ‘__uint32_t’ {aka ‘unsigned int’} from ‘int32_t’ {aka ‘int’} may change the sign of the result [-Wsign-conversion]
  258 |         resp.FTData[i] = static_cast<int32_t>(ntohl(resp4));
      |                                                     ^~~~~

+++ modules/sensor/src/force-torque/vpForceTorqueAtiNetFTSensor.cpp
@@ -255,7 +255,7 @@ bool vpForceTorqueAtiNetFTSensor::waitForNewData(unsigned int timeout)
       for (int i = 0; i < 6; i++) {
         int32_t resp4;
         memcpy(&resp4, &response[12 + i * 4], sizeof(int32_t));
-        resp.FTData[i] = static_cast<int32_t>(ntohl(resp4));
+        resp.FTData[i] = static_cast<int32_t>(ntohl(static_cast<uint32_t>(resp4)));
       }
       // Output the response data.
       if (resp.status) {
